Add canteens/:id/meals endpoint that days and meals.
Introduce new canteen_meals endpoint that returns up to 7 days
for given canteen and their meals. The new endpoint support start
filter like days endpoint. Output uses days and meals templates.

Intented use case is a single endpoint to fetch days and meals
for a canteen as well as day infos if canteen is closed or not
without the need to do two requests.

Changes:
* Add new day decorator to decorate day collection and day's meal
  association.
* Introduce FactoryGirl traits.
* Add route.