Fix deploy error: Missing sockets directory.
Failure on unicorn reload/start due to missing socket.
Directory tmp/sockets was not created because capistrano task
was not executed. It was registered as a before hook of
whenever:clear_crontab that seems to not be executed like before.

Fixed by registering deploy:create_sockets_symlink as a before
hook of deploy:create_symlink.