rake: fix om:fetch to freeing memory after each fetch
The rake task om:fetch receive all canteens and stores them in a array.
Afterwards the task iteraters above all canteens and fetches the current feed
data. Because of the top-level canteen list, all received or inserted meal
data (days, meals, notes) are not freed.

With the commit the task receives only the list of canteen ids and queries the
ActiveRecord only for the current canteen. I did not use find_bulk because it is
holding twice bulk size canteen in the memory.