Refactor anonymous user accessor
This refactors the `User::anonymous` method to not use a cached class
instance variable. Therefore not cleanup in e.g. tests is needed
anymore. The new method always looks up the record and uses a proper
table lock when creating an initial anonymous user. This avoids having
multiple anonymous users if the code runs parallel.