Switch from deprecated phantomjs to real browser
This CL removes the poltergeist driver using the deprecated PhantomJS
tool. It introduces support to run feature specs with a real browser,
chrome, using caprite, a CCD driver for capybara.

This CL starts to actually run all feature specs using chrome too,
instead of using the simulating rack-test driver. This means all tests
are run in an actual browser.

A specs needs to be adjusted to properly wait on the page for running
assertions or to account for whitespace in matches.