Allow (partial) blank values for near scope
has_scope 0.8.0 has a stricter validation for blank values in
combination with the `:using` option. Without allowing blank values, all
used fields must be present, otherwise the scope will not be applied.

The `:near` scoped for canteens expects that some fields are not set or
blank, it therefore needs the to have allow_blank set to True.