Remove the canteen wishlist feature
The wishlist feature allowed change public visible data without
moderation. As we have no capacity for manual moderation, in #27 it was
decided to remove this feature.

This CL removes the wishlist and removes all mentioned in other
features, such as selecting a wanted canteen when creating a feed
source. The canteen state "wanted" has been completely removed, and been
replaced by the "new" state, that is still used when creating new
canteens we sources.

Creating a new source has been reworked to directly show a canteen form
inline with the source form. Both records are created in a transaction.

Code and templates for the wishlist has been removed. Mentions in static
content and sidebars has been removed.

Closes #27.