Build a container image for the main application
This changeset is the result of the cummulative work of Christian and me
to build a container image to deploy and run the OpenMensa main
application.

We want to decouple the development and maintenance of the application
for operation requirements, such the available Ruby version on the
server and having to install the build and compile tools need on the
server. The application should be able to choose any Ruby version it
wishes and change the assets pipeline as need, for example, using NodeJS
to build some frontend assets without any changes to production.

For that, we add a CI pipeline and a container file definition, to build
the application. A multi-stage builds allows us to keep build
dependencies out from the final image.

Special considerations:

The image has been configured to copy all web assets into a mount volume
on boot, so that these can be served from an nginx container or an nginx
running on the host. The details for how we run that on the server still
needs to be finalized.

A docker-compose configuration is shipped for local testing, including
an nginx and postgres. This might ease in getting a much more
production-like setup for e.g. new contributors.

Testing:

The container image is continously built and published in CI using
Github Actions, for every branch. We might need to manually delete not
used tags from old branches every now and then. We did have many
branches in the last years at all, so it shouldn't be a big problem.

The image itself isn't used to run any tests, nor is it tested itself.
We might want to add that as soon as we know how we will use the image,
to ensure it is actually working.

Deployment:

We actually do not use the image in production yet, and havn't finalized
how we want to configure it. Currently, the correct configuration files
need to be mounted.