docs: Fix repo name + edit button links to source
On GitHub and GitLab, the default "edit" path (edit/master/docs/) opens the
page in the online editor. This functionality requires that the user have
and be logged in to a GitHub/GitLab account. Otherwise, the user will be
redirected to a login/signup page.

Instead, we will link to the blob page.