fix: N+1 query in canteen_meals endpoint
The MealDecorator overrode #decorate_collection to explicitly add an
`includes(:notes)` but that interferes with associations already having
a matching preload, such as the canteen_meals endpoint that preloads
meals+notes on the days collection already. This resulted in all meals
and notes being loaded once with the days, and again for each day in the
MealDecorator#decorate_collection.

This CL removes the #decorate_collection override to fix the N+1 query
problem for canteen_meals. It further adds some new Rails strict_loading
scoping to these endpoints, and adds an explicit preload for meal notes
for MealsController#index through the default_scope defined there.

Fixes OPENMENSA-21
Fixes OPENMENSA-1Y