feat: Upgrade to Rails 7.2
Replace the removed Rails.secret framework with rubyconfig, a config
file framework with support for validation.

This includes custom code to read the Rails secret key base from that
config file, instead of Rails credentials, since we DO NOT store any
production credentials in a repository here. All production
configuration is pushed on deployment, and it does not make any sense to
push an encrypted credentials files and a master key from the same
source to the same service.

This repository is an the application source, and does not contain any
credentials. As far as possible, code can be deployed or rolled backed,
without affecting deployed configuration.